Made EnemySpawner_Eric
EnemySpawner_Eric does the same purpose as EnemySpawner. But
EnemySpawner_Eric can be repeatedly used, and uses bezierSplineFollower
script to create the enemy with a curve at runtime instead of spawning a
prefab with the curve built in. This curve will most likely be changed
anyway if any AI is to have a non-bezier curve applie to it, but for
right now, it'll do.